Refactor mission button
